nest js

+++ src/Modules/tasks/services/tasks.service.ts
@@ -27,7 +27,7 @@ export class TasksService {
     });
   }
 
-  async getOne(id: number): Promise<Task> {
+  async getOne(id: number, userId: number): Promise<Task> {
     const task = await this.taskModel.findByPk(id, {
       include: [
         {
@@ -41,6 +41,9 @@ export class TasksService {
     if (!task) {
       throw new CustomError(taskNotFound, 400);
     }
+    if (task.userId !== userId) {
+      throw new CustomError('You are not authorized', 401);
+    }
     return task;
   }
 
@@ -53,10 +56,17 @@ export class TasksService {
       userId,
     });
   }
-  async update(id: number, taskUpdate: TaskCreateDTO): Promise<any> {
+  async update(
+    id: number,
+    taskUpdate: TaskCreateDTO,
+    userID: number,
+  ): Promise<any> {
     const task = await this.taskModel.findByPk(id);
     if (!task) {
       throw new CustomError(taskNotFound, 400);
+    }
+    if (task.userId !== userID) {
+      throw new CustomError('You are not authorized', 401);
     }
     const { title, description, status, userId } = taskUpdate;
     return await this.taskModel.update(
@@ -71,10 +81,13 @@ export class TasksService {
       },
     );
   }
-  async delete(id: number): Promise<any> {
+  async delete(id: number, userID: number): Promise<any> {
     const task = await this.taskModel.findByPk(id);
     if (!task) {
       throw new CustomError(taskNotFound, 400);
+    }
+    if (task.userId !== userID) {
+      throw new CustomError('You are not authorized', 401);
     }
     return await this.taskModel.destroy({
       where: { id },
